@@ -1,10 +1,10 @@
 name: flutter_scale_tap
 description: Scale and opacity animated Button for Flutter
-version: 1.0.4
+version: 1.0.5
 repository: https://github.com/fedeamura/flutter_scale_tap
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
